+++ custom_components/hacs/frontend/hacs.css
@@ -11,3 +11,12 @@ a {
   color: #ffab40;
 }
 
+.errorview {
+  display: block;
+  margin-bottom: 30px;
+}
+
+.hacsbutton {
+  margin-left: 3%;
+  margin-top: 5%;
+}
