+++ custom_components/hacs/frontend/views/__init__.py
@@ -1,9 +1,4 @@
 """Initialize frontend views."""
-from custom_components.hacs.frontend.views.error import error_view
-from custom_components.hacs.frontend.views.api import CommunityAPI
-from custom_components.hacs.frontend.views.element import CommunityElement
-from custom_components.hacs.frontend.views.overview import CommunityOverview
-from custom_components.hacs.frontend.views.plugin import CommunityPlugin
-from custom_components.hacs.frontend.views.settings import CommunitySettings
-from custom_components.hacs.frontend.views.store import CommunityStore
+from custom_components.hacs.frontend.views.error import HacsErrorView
 from custom_components.hacs.frontend.views.static import HacsStaticView
+from custom_components.hacs.frontend.views.plugin import HacsPluginView
