--- custom_components/hacs/frontend/views/settings.py
@@ -1,188 +1,0 @@
-"""CommunityAPI View for HACS."""
-import logging
-from aiohttp import web
-from homeassistant.components.http import HomeAssistantView
-from homeassistant.const import __version__ as HAVERSION
-
-from custom_components.hacs.const import DOMAIN_DATA, NAME_LONG
-from custom_components.hacs.frontend.views import error_view
-from custom_components.hacs.frontend.elements import (
-    info_card,
-    style,
-    header,
-    generic_button_external,
-    generic_button_local,
-    warning_card,
-)
-
-_LOGGER = logging.getLogger('custom_components.hacs')
-
-
-class CommunitySettings(HomeAssistantView):
-    """View to serve the overview."""
-
-    requires_auth = False
-
-    url = r"/community_settings"
-    name = "community_settings"
-
-    def __init__(self, hass, hacs):
-        """Initialize overview."""
-        self.hass = hass
-        self.hacs = hacs
-        self.message = None
-
-    async def get(self, request):  # pylint: disable=unused-argument
-        """View to serve the overview."""
-        _LOGGER.debug("Trying to serve settings")
-        self.message = request.rel_url.query.get("message")
-        try:
-            html = await self.settings_view()
-        except Exception as error:  # pylint: disable=broad-except
-            _LOGGER.debug(error)
-            html = await error_view()
-        return web.Response(body=html, content_type="text/html", charset="utf-8")
-
-    async def settings_view(self):
-        """Settings view."""
-        content = ""
-        content += await style()
-        content += await header(self.hacs)
-
-        content += "<div class='container'>"
-
-        if self.hacs.data["hacs"].get("restart_pending"):
-            content += await warning_card(
-                "You need to restart Home Assisant to start using the latest version of HACS."
-            )
-
-        if (
-            self.hacs.data["hacs"]["local"]
-            != self.hacs.data["hacs"]["remote"]
-        ):
-            content += """
-                <div class="row">
-                    <div class="col s12">
-                    <div class="card  red darken-4">
-                        <div class="card-content white-text">
-                        <span class="card-title">UPDATE PENDING</span>
-                        <p>There is an update pending for HACS!.</p>
-                        </br>
-                        <p><b>Current version:</b> {local}</p>
-                        <p><b>Available version:</b> {remote}</p>
-                        </div>
-                        <div class="card-action">
-                        <a href="/community_api/hacs/upgrade"
-                            onclick="document.getElementById('progressbar').style.display = 'block'">
-                            UPGRADE</a>
-                        <a href="https://github.com/custom-components/hacs/releases/tag/{remote}" target="_blank">
-                            CHANGELOG</a>
-                        </div>
-                    </div>
-                    </div>
-                </div>
-            """.format(
-                local=self.hacs.data["hacs"]["local"],
-                remote=self.hacs.data["hacs"]["remote"],
-            )
-
-        # Show info message
-        if self.message is not None and self.message != "None":
-            content += await warning_card(self.message)
-
-        # Integration URL's
-        content += """
-        <div class="row">
-                <ul class="collection with-header">
-                    <li class="collection-header"><h5>CUSTOM INTEGRATION REPO'S</h5></li>
-        """
-        if self.hacs.data["custom"].get("integration"):
-            for entry in self.hacs.data["repos"].get("integration"):
-                content += """
-                    <li class="collection-item">
-                        <div><a href="/community_api/integration_url_reload/{url}" onclick="document.getElementById('progressbar').style.display = 'block'"><i class="fa fa-sync" style="color: #26a69a; margin-right: 1%"></i></a>  {entry}
-                            <a href="/community_api/integration_url_delete/{url}" class="secondary-content">
-                                <i name="delete" class="fas fa-trash-alt"></i>
-                            </a>
-                        </div>
-                    </li>
-                """.format(
-                    entry=entry, url=entry.replace("/", "%2F")
-                )
-        content += """
-                </ul>
-            <form action="/community_api/integration_url/add" method="post" accept-charset="utf-8" enctype="application/x-www-form-urlencoded">
-                <input id="custom_url" type="text" name="custom_url" placeholder="ADD CUSTOM INTEGRATION REPO" style="width: 90%">
-                    <button class="btn waves-effect waves-light right" type="submit" name="add" onclick="document.getElementById('progressbar').style.display = 'block'">
-                        <i class="fas fa-save"></i>
-                    </button>
-            </form> 
-            </br>
-        """
-
-        # Plugin URL's
-        content += """
-                <ul class="collection with-header">
-                    <li class="collection-header"><h5>CUSTOM PLUGIN REPO'S</h5></li>
-        """
-        if self.hacs.data["custom"].get("plugin"):
-            for entry in self.hacs.data["repos"].get("plugin"):
-                content += """
-                    <li class="collection-item">
-                        <div><a href="/community_api/plugin_url_reload/{url}" onclick="document.getElementById('progressbar').style.display = 'block'"><i class="fa fa-sync" style="color: #26a69a; margin-right: 1%"></i></a>  {entry}
-                            <a href="/community_api/plugin_url_delete/{url}" class="secondary-content">
-                                <i name="delete" class="fas fa-trash-alt"></i>
-                            </a>
-                        </div>
-                    </li>
-                """.format(
-                    entry=entry, url=entry.replace("/", "%2F")
-                )
-        content += """
-                </ul>
-            <form action="/community_api/plugin_url/add" method="post" accept-charset="utf-8" enctype="application/x-www-form-urlencoded">
-                <input id="custom_url" type="text" name="custom_url" placeholder="ADD CUSTOM PLUGIN REPO" style="width: 90%">
-                    <button class="btn waves-effect waves-light right" type="submit" name="add" onclick="document.getElementById('progressbar').style.display = 'block'">
-                        <i class="fas fa-save"></i>
-                    </button>
-            </form> 
-        </div>
-        """
-
-        content += "</br>"
-        content += "</br>"
-        content += await generic_button_local(
-            "/community_api/self/reload", "RELOAD DATA"
-        )
-        content += await generic_button_external(
-            "https://github.com/custom-components/hacs/issues/new/choose", "OPEN ISSUE"
-        )
-        content += await generic_button_external(
-            "https://github.com/custom-components/hacs", "HACS REPO"
-        )
-        content += await generic_button_external("/community_api/log/get", "OPEN LOG")
-        content += "</br>"
-        content += "</br>"
-        if self.hacs.data["hacs"].get("restart_pending"):
-            local_version = "{} <b>(RESTART PENDING!)</b>".format(
-                self.hacs.data["hacs"]["local"]
-            )
-        else:
-            local_version = "{}".format(self.hacs.data["hacs"]["local"])
-        info_message = """
-        <h5>{}</h5>
-        <b>HACS version:</b> {}</br>
-        <b>Home Assistant version:</b> {}</br>
-        </br>
-        <hr>
-        <h6>UI built with elements from:</h6>
-        <li><a href="https://materializecss.com" target="_blank" style="font-weight: 700;">Materialize</a></li>
-        <li><a href="https://fontawesome.com" target="_blank" style=";font-weight: 700;">Font Awesome</a></li>
-        <hr>
-        <i>This site and the items here is not created, developed, affiliated, supported, maintained or endorsed by Home Assistant.</i>
-        """.format(
-            NAME_LONG, local_version, HAVERSION
-        )
-        content += await info_card(info_message)
-        content += "</div>"  # End the view container
-        return content
