+++ custom_components/hacs/hacsviewbase.py
@@ -24,9 +24,9 @@ class HacsViewBase(HomeAssistantView, HacsBase):
         return f"""
         <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/materialize/1.0.0/css/materialize.min.css">
         <link rel="stylesheet" href="https://use.fontawesome.com/releases/v5.8.2/css/all.css" integrity="sha384-oS3vJWv+0UjzBfQzYUhtDYW+Pj2yciDJxpsK1OYPAYjqT085Qq/1cq5FLXAZQ7Ay" crossorigin="anonymous">
-        <script src="https://cdnjs.cloudflare.com/ajax/libs/materialize/1.0.0/js/materialize.min.js"/>
+        <script src="https://cdnjs.cloudflare.com/ajax/libs/materialize/1.0.0/js/materialize.min.js"/></script>
         <link rel="stylesheet" href="{self.url_path["static"]}/hacs.css">
-        <script src={self.url_path["static"]}/hacs.js"/>
+        <script src={self.url_path["static"]}/hacs.js"></script>
         """
 
     @property
